requirements.txt
alabaster==0.7.11
appnope==0.1.0
asn1crypto==0.24.0
astroid==1.6.5
autopep8==1.3.5
Babel==2.6.0
backcall==0.1.0
bleach==2.1.3
cachetools==2.1.0
certifi==2018.8.13
cffi==1.11.5
chardet==3.0.4
colorama==0.3.9
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.3.1
decorator==4.3.0
Django==2.1
django-cors-headers==2.4.0
django-cors-middleware==1.3.1
django-crispy-forms==1.7.2
django-filter==2.0.0
django-filters==0.2.1
django-guardian==1.4.9
django-rest-swagger==2.2.0
djangorestframework==3.8.2
docutils==0.14
entrypoints==0.2.3
google-api-python-client==1.7.4
google-auth==1.5.0
google-auth-httplib2==0.0.3
html5lib==1.0.1
httplib2==0.11.3
idna==2.7
imagesize==1.0.0
inapppy==0.6
ipykernel==4.8.2
ipyparallel==6.2.1
ipython==6.4.0
ipython-genutils==0.2.0
ipywidgets==7.2.1
isort==4.3.4
itypes==1.1.0
jedi==0.12.0
Jinja2==2.10
jsonschema==2.6.0
jupyter-client==5.2.3
jupyter-core==4.4.0
lazy-object-proxy==1.3.1
Markdown==2.6.11
MarkupSafe==1.0
mccabe==0.6.1
mistune==0.8.3
mysqlclient==1.3.13
nbconvert==5.3.1
nbformat==4.4.0
notebook==5.5.0
oauth2client==4.1.2
openapi-codec==1.3.2
pandocfilters==1.4.2
parso==0.2.1
pickleshare==0.7.4
Pillow==5.2.0
polyline==1.3.2
prompt-toolkit==1.0.15
pyasn1==0.4.4
pyasn1-modules==0.2.2
pycodestyle==2.4.0
pycparser==2.18
Pygments==2.2.0
PyMySQL==0.9.2
PyPDF2==1.26.0
python-dateutil==2.7.3
pytz==2018.5
pywinpty==0.5.4
pyzmq==17.0.0
reportlab==3.5.6
requests==2.19.1
rsa==3.4.2
Send2Trash==1.5.0
simplegeneric==0.8.1
simplejson==3.16.0
six==1.11.0
terminado==0.8.1
testpath==0.3.1
tornado==5.0.2
traitlets==4.3.2
uritemplate==3.0.0
urllib3==1.23
wcwidth==0.1.7
webencodings==0.5.1
widgetsnbextension==3.2.1
wrapt==1.10.11
xhtml2pdf==0.2.2
